requirements.txt
requests>=2.25.0
websocket-client>=1.8.0
Pillow
numpy
scikit-learn
